requirements.txt
flask
requests
python-dotenv
matplotlib
plotly
pandas